+++ spec/philiprehberger/id_gen_spec.rb
@@ -62,12 +62,43 @@ RSpec.describe Philiprehberger::IdGen do
     it 'accepts worker_id keyword' do
       expect(described_class.snowflake(worker_id: 1)).to be_a(Integer)
     end
+
+    it 'accepts a custom epoch' do
+      custom_epoch = Time.utc(2015, 1, 1)
+      id = described_class.snowflake(worker_id: 0, epoch: custom_epoch)
+      expect(id).to be_a(Integer)
+      expect(id).to be > 0
+    end
   end
 
   describe '.snowflake_timestamp' do
     it 'extracts a Time from a snowflake ID' do
       id = described_class.snowflake
       result = described_class.snowflake_timestamp(id)
+      expect(result).to be_a(Time)
+      expect(result).to be_within(1).of(Time.now)
+    end
+
+    it 'extracts a Time using a custom epoch' do
+      custom_epoch = Time.utc(2015, 1, 1)
+      id = described_class.snowflake(worker_id: 0, epoch: custom_epoch)
+      result = described_class.snowflake_timestamp(id, epoch: custom_epoch)
+      expect(result).to be_a(Time)
+      expect(result).to be_within(1).of(Time.now)
+    end
+  end
+
+  describe '.uuid_v7' do
+    it 'returns a valid UUID v7 string' do
+      uuid = described_class.uuid_v7
+      expect(uuid).to match(/\A[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\z/)
+    end
+  end
+
+  describe '.uuid_v7_timestamp' do
+    it 'extracts a Time from a UUID v7' do
+      uuid = described_class.uuid_v7
+      result = described_class.uuid_v7_timestamp(uuid)
       expect(result).to be_a(Time)
       expect(result).to be_within(1).of(Time.now)
     end